On-Disk File System
The old xv6 file system is supported on disk. In addition an
'fs_log' is created that maintains a log of all file system
operations and flushes them out to disk periodically. Currently
this is supported via the command line tool 'fsflush' that writes
to disk the file system operations logged after the last call to
fsflush. Also, in the current version there is just one FS
operations log that is shared over all the cores.