mfs_logical_log: a logical log of filesystem metadata updates
Metadata operations that take place on the in-memory representation of the
filesystem are stored in the logical log. When fsync() is called on a
file/directory, the operations in this log are first applied to the disk
filesystem to take care of any dependencies, and then the fsync goes
through. The logical metadata log is implemented using oplog.