No need to track newly created files/directories
There won't be any orphan files that need to be freed up during system
recovery. The file create will have been logged in the logical metadata log
and will get executed before an fsync is called on the file. In the current
design, a file create also creates a corresponding directory entry in the
parent directory. The same reasoning holds for new directories as well.