Crash recovery: persistent physical log
The physical log is written to a journal on the disk. During boot, committed
transactions in this journal are replayed. Each transaction is made up of a
start block, a sequence of disk blocks (corresponding to the changes made) and
a commit block. The journal only contains transactions that were logged but not
applied to the disk filesystem, not the ones that made it safely to the disk.
NOTE: The journal is now implemented using the file "sv6journal". mkfs needs to
be modified to allocate a fixed inode to the journal file, the number of which
can perhaps then be stored in the superblock.