Get tsc values at linearization points
To timestamp metadata operations at their linearization points, we use the tsc
value obtained when the mdir's hash table is updated under a lock. All of these
updates (insert, remove, rename) are atomic, so all this involves is a call to
rdtsc() within the locked region. The special case where the source and
destination for a rename are the same has not been handled yet.