Wait for in-flight operations to complete before synchronizing oplog loggers
When the per-core metadata logs are synchronized during an fsync() call, there
might be some operations with linearization points before the fsync call, but
which have not been logged yet. We need to wait for these operations to be
logged in order to not miss them. Each core updates a tsc counter indicating the
starting and ending times of metadata operations. These counters are used to
figure out if there are operations executing on some cores which might have
linearization points before the fsync call.