Check for dependencies and apply only relevant metadata operations
In the previous design an fsync() flushed out the entire metadata log. This
would incur a whole lot of independent writes to the disk and hamper
performance. For example, if two distinct files are created, and we fsync just
one of them, the other file creation would be written to disk as well. Instead
we now traverse the metadata log and apply only those operations that the
fsync() depends on. So we only write to disk as much as is necessary.