scalefs: Fix unnecessary copying in add_block
Originally, this code simply added the passed pointer to blocks.  The
conversion to unique_ptr was safe but not equivalent: the use of
make_unique allocated a *new* transaction_diskblock, and then moved
the passed transaction_diskblock into the new one (which is an
expensive operation because transaction_diskblock is big), and then
added wrapped that in a unique_ptr.  Instead, the intent was simply to
wrap the existing pointer in a unique_ptr.