scalefs: Even better, take unique_ptr in add_block
Rather than taking a bare pointer and converting it to a unique_ptr,
take a unique_ptr in add_block and add_blocks.  This way, the entire
lifetime of the transaction_diskblock is automatically managed.  This
also eliminates a bunch of std::move(new X) (applying std::move to a
pointer is effectively a no-op, since moving and copying a pointer are
no different).