kalloc: Don't crash when a NUMA node has no CPUs
Previously, initkalloc would crash with a divide-by-zero when ACPI
reported a NUMA node but no associated CPUs because it would attempt
to divvy the node's memory across its CPUs.  I'm not sure this can
ever happen on real hardware, but it's easy to trigger on QEMU if you
set the CPU count to 1, but leave the default two NUMA nodes.

Fix this by ensuring every NUMA node is sub-divided into at least 1
subnode.