mfs, boot: Improve booting by allocating memory for free_bit_vector in bulk
It turns out that optimizing the memory allocation for the free_bit_vector
can give speed improvements of upto 650% and reduce memory consumption by
nearly half!

During boot, we read the free-bit bitmaps of the filesystem from the disk
and initialize an in-memory data-structure called 'free_bit_vector' to
maintain a copy of those bitmaps. Since the disk is relatively small (with
32768 blocks), all the free-bits fit into just a single disk block. So the
disk read is actually inexpensive, since we only read one single block.

However, what follows after the disk read, is a loop that runs 32768 times,
in which the free_bit_vector is repeatedly expanded to accommodate info
about all those free bits. Each free-bit object is 256 bytes, so the total
memory needed is exactly 8 MB. However, the kernel ends up consuming nearly
32 MB for just this, which is nearly 4 times the required amount!

Instead if we count the number of free-bits and pre-allocate the memory for
the free_bit_vector in one shot, it consumes only 16 MB of memory. Also,
since we avoid invoking the memory allocator repeatedly, it reduces the time
for this loop by over 6 times as shown below.

Without patch:
--------------

Time taken      :  758 ms
Memory consumed : 8190 pages (~ 32 MB)


With patch:
-----------

Time taken      :  116 ms
Memory consumed : 4096 pages (~ 16 MB)


Going by the absolute numbers, the speed improvement is not noticeable,
whereas the space improvement shows that the behavior of the page allocator
is quite worrisome - it handles small-size allocations very poorly!
(This behavior also extends to allocations that cross a page by a few bytes,
as confirmed by other experiments).

While this patch manages to restore 16 MB of the system memory back to
userspace, it is worthwhile to take a deeper look at optimizing the page
allocator's space-efficiency for small-size requests, to prevent squandering
memory.