tools/mkfs: Reduce the size of fs.img to 16 MB, to prepare for upcoming changes
PXE booting fails when the size of the kernel image is large, particularly
when we bake the disk image directly into the kernel image (MEMIDE=1). A disk
size of 64 MB works fine, but larger sizes just don't boot. So we are in the
process of reworking this whole process so as to use compression to solve the
problem. To facilitate this change, reduce the size of fs.img to 16 MB so as to
mimic the master branch.