tools: Import the zlib compression library into sv6 to help with PXE booting
When MEMIDE is set to 1, we bake the disk image directly into the kernel image
and read/write the main-memory to simulate disk accesses. Unfortunately, PXE
booting fails when the size of the kernel image is relatively large. (For ex:
it works when the filesystem (disk) size is 64 MB, but fails when it reaches
128 MB or higher).

To workaround this limitation, import the zlib compression library, so that
we can compress the filesystem image during build and then decompress it back
in memory during boot, thereby allowing us to work with larger disk sizes.

Importing the zlib source code is mostly straight-forward, except for a few
changes noted below:

1. In the zlib source code, there is a dummy Makefile with these contents:

all:
        -@echo "Please use ./configure first.  Thank you."

distclean:
        make -f Makefile.in distclean

This Makefile gets overwritten when we run ./configure in the zlib directory.
But if this Makefile is already tracked by git (as part of the zlib source
code), it keeps tracking the updates to this file every time we build the
source, which would interfere with tracking the actual source code changes
which we would like to make during development.

We avoid this problem by running ./configure --static and generating the
final Makefile before committing the zlib source code to the sv6 repository.

2. We would like to build libz.a and then link it with the sv6 kernel image.
But the default Makefile configuration in zlib leads to linker errors while
attempting this (eg: relocation truncated to fit: R_X86_64_32 against `.rodata').
To fix this, we modify the generated Makefile of zlib, so as to pass the
parameter "-mcmodel=kernel" to the compiler.

Also, while we are at it, copy the zlib.h and zconf.h headers from zlib to
the include/ directory of sv6, as we'll need it eventually for writing the
in-kernel decompression code.