boot, zlib: Decompress fs.imgz in memory during boot
During boot, we need to extract the full disk image from the compressed
version (fs.imgz) embedded within the kernel image. So alter the memide code
to decompress fs.imgz to a statically allocated array in memory.

Unfortunately, we can't dynamically allocate contiguous chunks of memory in
the kernel larger than 16 MB, due to the limitations of the buddy allocator.
So we use static allocation, and hence depend on knowing the (uncompressed)
disk size at compile-time.

And now that we have made the necessary changes to the memide code to lookup
fs.imgz in the kernel image instead of fs.img, we can safely exclude fs.img
from the kernel image.