scalefs: Remove incorrect call to release() in mfs_interface::find_free_block()
Commit 0c35cd569 (scalefs: Fix locking bug in mfs_interface::find_free_block())
added an explicit call to release() while trying to fix the locking issue in
find_free_block(). But this is buggy, because the 'lock' going out of scope
automatically releases the write_lock. (Also, since the commit invokes
it->write_lock.release(), we end up trying to release a lock that is not even
held!). Fix this by removing the explicit call to release().