linearhash: Destroy the value (object) during remove()
In the current implementation of the linear hash-table, remove() simply sets
the valid flag to false, but doesn't actually destroy the value. This works
fine for simple <key, value> pairs where the values are just integers.

However, when using the hash-table to store more complicated values such as
sref objects, we should also explicitly destroy the object during remove(),
because otherwise, the decrement of the object's reference-count will get
missed. (The corresponding increment would have happened when the value
got inserted into the hash-table).

So, fix this by destroying the object during remove().