fs: Don't increment the inode link count in ialloc()
Calls to ialloc() are followed by invocations of dirlink(), which
increments the inode's link count. So, doing an additional increment
in ialloc() itself leads to an extra link count that never gets dropped
(and as a result, unlinking a file doesn't release its inode). Fix this
by removing the increment in ialloc().