scalefs: Update the free-bit-vector in flush_journal()
After committing a transaction to the journal and applying it to the
filesystem, we should also mark the freshly freed blocks as free in the
in-memory free-bit-vector.

add_fsync_to_journal() remembers to perform this update, but flush_journal()
misses it. As a result, the free-bit-vector can show stale values, and more
importantly, since we consult this vector for block allocation, this bug can
leak disk blocks.

For example, when we unlink a file and fsync the file's parent directory,
flush_journal() is invoked, and due to this bug, that file's disk blocks will
never get reused for any further allocations. Fix this by updating the
free-bit-vector in flush_journal().