fs: Don't log the zeroing of data blocks in the fixed-size journal
When unlinking a file and invoking fsync on its parent directory, the file
is truncated and its blocks are freed by calling bfree(). bfree() will
internally zero out those blocks before noting that they are free. However,
this zeroing operation is performed using bzero(), which *logs* the updated
disk blocks in the transaction.

According to commit 506cc669 (Make the on-disk journal static), the journal
is currently intended to be a static file of fixed size that never grows.
Obviously its size is limited, and as per design, we want to log only the
changes to the metadata blocks in the journal and not the updates to the
actual data blocks.

Since bzero() logs everything in the journal, unlinking large files overflows
the journal and leads to an unintended expansion of the journal file
(/sv6journal), making it appear as if we lost disk blocks mysteriously.

Fix this issue by being judicious about logging zeroed out blocks - only
log the metadata blocks that are being zeroed and freed, not the data blocks
themselves. Zero out the data blocks directly on the disk.

Introduce suitable helpers named bfree_metadata() and bfree_datablock() to
achieve this.