fs: Don't zero out blocks when freeing if they have not been written to
In bmap(), we allocate free blocks on-demand and try to link them to the
file's metadata. If that fails, we retry the operation after freeing the
block that was just allocated. At this point, those blocks have not been
touched at all, so there is no point zeroing them out while freeing.

So add a variant of bfree() called bfree_nozero() that skips the zeroing
part, and invoke it from bmap().