fs, bmap: Fix error recovery in indirect-block-allocation failure path
In bmap(), if the attempt to atomically assign the newly allocated indirect
block to ip->iaddrs fails, we free the allocated block and the memory
corresponding to iaddrs, but we don't erase the entry for that allocated
block that we had already put in ip->addrs[NDIRECT]. But since that block
('addr') was just freed in the failure path, ip->addrs[NDIRECT] will end up
referring to a free block!

Fix this by resetting ip->addrs[NDIRECT] to 0 in this failure path.