fs, scalefs: Document the size of the physical journal
At the moment, the size of /sv6journal is 4235264 bytes, which is 1034 blocks
(assuming BSIZE=4096). This happens to be the maximum size of a file consisting
only of direct and indirect blocks. However, there are situations (large
transactions, filesystem configuration etc) that might require more space in
the journal. We won't increase the size of the journal just yet (since it is
large enough for the moment), but let us write down some of the calculations
that go into determining this size, should we need to increase it later.

As for the code in this commit, demystify the magic number 4235264!