scalefs: Initialize the free-bit-vector *after* applying committed transactions
While initializing scalefs during boot, we invoke initialize_free_bit_vector(),
which reads the free bitmap blocks of the filesystem from the disk and
constructs the in-memory free-bit-vector. However, this is done a little too
early, even before processing the journal and applying the committed
transactions.

Since transactions can include updates to the free bitmap blocks, the in-memory
free-bit-vector can get totally out-of-sync with the actual state on the disk!
For example, if we crash just after committing a transaction that allocates a
few disk blocks, upon reboot, even though the transaction is processed and
applied, the in-memory free-bit-vector can end up indicating that those blocks
are free and reuse them for other purposes!

Fix this by initializing the free-bit-vector after processing the journal.