mm, buddy: Keep a record of the amount of free memory at its lowest
While characterizing the memory overhead of ScaleFS, one of the interesting
parameters that we would like to observe is the maximum amount of kernel
memory that is ever consumed. This gives us an idea about the worst-case
behavior of the kernel in terms of memory consumption.

To observe this, keep a record of the lowest free memory in the buddy
allocator and print it out along with the rest of the memory statistics.