scalefs: Save memory by tracking only unique diskblocks in a transaction
Currently, we keep track of all updates to metadata blocks in a transaction,
including multiple updates to the same set of diskblocks. This turns out to
be very expensive (in terms of the memory consumed), especially when some
of the metadata blocks are updated heavily within a transaction.

For example, consider creating a file and writing a large amount of data to
it and then invoking fsync on it. Once we start growing the file, soon we'll
end up having to update the indirect and the doubly-indirect blocks (which
are logged in the transaction, since they are metadata updates). After this
point, every block allocation for that file ends up touching the indirect or
the doubly-indirect blocks, thereby adding close to 1x memory overhead to the
transaction!

Unfortunately, there are several more situations as well, where we have this
problem, and most of them are not trivial to avoid without a substantial
redesign. So instead, solve this at the other end: add a hash-table to the
transaction to maintain a list of the currently logged blocks, and use it to
efficiently detect updates to the same block. When this is detected, just
update the contents of the existing transaction_diskblock instead of creating
a new one. Thus we can ensure that the transaction only tracks unique
diskblocks, all of which have the latest contents.

[Important note: We already manage the blocks in the transaction-diskblock
vector using unique_ptr. We would like to continue to manage memory using
those pointers alone. The hash-table serves an entirely different purpose,
so we just use raw (non-owning) pointers to these blocks in the hash-table.]

This enhancement reduces the memory overhead of fsync() by nearly 1x, as
shown below:

Experiment: Boot the machine with 512 MB RAM, create a file and write 30 MB
            of data and invoke fsync on that file.

Before patch:
-------------

Free memory at test start                : 105072 pages
Lowest recorded free mem after fsync     :  78290 pages
Ratio of memory consumed to data written : (105072 - 78290) / 7680 = 3.49


After patch:
------------

Free memory at test start                : 105072 pages
Lowest recorded free mem after fsync     :  85648 pages
Ratio of memory consumed to data written : (105072 - 85648) / 7680 = 2.53

Thus this patch reduces the run-time memory consumption of fsync by 1x.