scalefs: Update the free-bit-vector as soon as the transaction commits
Blocks that are freed due to operations such as truncating or deleting a
file should not be available for reuse until such transactions get committed
to the disk. This is the reason why the freed blocks are not immediately
marked as free in the in-memory free-bit-vector.

However, the current code delays this update for too long, not only until
the transaction is committed to the journal, but also until it is applied
to the filesystem and furthermore, until the journal is cleared!

Fix this by performing the update much earlier, immediately after the
transaction commits (because that's all we really need to wait for).

As a simple rule-of-thumb, never prolong granting access to resources (disk
blocks in this case) without a legitimate reason, lest the system build up
unnecessary pressure for that resource, causing unforeseen bottlenecks!