strbuf: Print a warning message when path components are longer than DIRSIZ
DIRSIZ is rather tiny (14 characters) and hence it is not uncommon to
inadvertently cross that limit once in a while. However, it is unfortunate
that this leads to weird behavior and it takes quite a bit of debugging to
figure out what actually went wrong (mainly because we return -1 as the error
code for pretty much all error scenarios).

The right fix is to encode separate error return values for different errors.
But for the time being, let us at least print a warning to demystify what
went wrong.

(Also, it might not hurt to think of relaxing DIRSIZ to a larger value in
the future, after taking care of any dependencies).