buf: Introduce add_to_transaction() to copy dirty bufcache block to a transaction
There are a lot of places where we make changes to a buffer-cache block, and
then include a copy of that updated block in a transaction, like this:

    auto locked = bp->write();

    //modify locked->data

    char charbuf[BSIZE];
    memmove(charbuf, locked->data, BSIZE);
    trans->add_unique_block(blocknum, charbuf);

Introduce add_to_transaction() to encapsulate this common usage; and while at it,
avoid the rather unnecessary copy to charbuf. (Avoiding this copy has another
important benefit: the size of this on-stack buffer is BSIZE, and hence if that
value increases, there is a good chance that the kernel stack will explode!).