buf: Introduce write_clean() to get exclusive (write) access to the block
Today, we have the option of taking either the read or the write lock over
the bufcache blocks. But the write lock (using ->write()) automatically
marks the block as dirty, assuming that we are going to update the contents
of the block.

However, there are situations where we don't intend to change the contents
of the block itself, but would like to update its other aspects such as
reference counts. In such cases, we still need the strongest synchronization
scheme (equivalent to ->write(), to prevent racing with concurrent writers),
but without the automatic updates to the dirty flag and the reference count.

So introduce ->write_clean() as a way to get exclusive access to the block
for such scenarios.