buf: Introduce buf::put() to evict blocks from the buffer-cache
We have buf::get() to load blocks into the buffer-cache. In the future,
we would like to have a mechanism to evict selected blocks from the
buffer-cache. To achieve this, introduce the counterpart of buf::get()
called buf::put(), having a similar interface.

Since bufcache blocks are managed using reference counts, all buf::put()
has to do is drop the extra reference count that was added in buf::get().
This will ensure that the block will get evicted from the cache and its
buffer will get freed up.

However, we need to take care to get exclusive access to the block when
manipulating the reference count, to avoid racing with concurrent writers
(we don't want to free up the buffer from under a writer who is about to
update its contents!). For this, we need to use ->write_clean() rather
than ->write(), so that we don't make unintended updates to the dirty-flag
or the reference count. Also, make sure to evict only clean blocks from the
buffer-cache.