buf: Introduce evict_bufcache() to evict blocks from the buffer-cache
Add evict_bufcache() to go through each file in the filesystem that has been
referenced (by looking it up in the MFS layer) and drop all the cached blocks
of those files. The helper named drop_bufcache() performs the actual task of
invoking buf::put() on all the blocks of the file, including the metadata blocks.

This approach/interface of evicting blocks at a file-level has several
advantages over, say, iterating through all the blocks in the buffer-cache and
evicting all the clean blocks blindly. Having a file-level granularity is not
only advantageous on its own, but it also helps us to easily avoid kicking out
some rather important (heavily-used) blocks from the bufcache, such as the
free-bitmap blocks, inode blocks, blocks containing directory entries and so on.

Another notable aspect is that this bufcache eviction is kept independent from
sync/fsync, so that we can play around with them separately.