fs: Fix create(), to allow opening device files using the shell's '>' operator
We should be able to open and write to device files using a command such as:

echo "hello" > /dev/device_file

However, this doesn't work at the moment because, upon seeing the redirection
operator '>', the shell tries to open the file with O_CREAT (along with other
flags). This leads to the create() function in the filesystem, which successfully
finds 'device_file' upon a lookup, but refuses to proceed further for anything
other than a regular file. Fix this issue by teaching create() to include
device files in this category.