buf: Call add_to_transaction() with the buf's write_lock_ held
Whenever we update a disk block and add it to a transaction, the expectation
is that *that* particular version of the block is added to the transaction.
Hence, we cannot afford to drop the write_lock_ between updating the block
and adding it to the transaction, lest some other concurrent writer update
the block's contents in the meantime!

However, we could potentially copy the contents of the block to a temporary
buffer with the write_lock_ held, and then copy this buffer to the transaction
at leisure, with the write_lock_ dropped. It turns out that we can't really
do this due to another subtle reason - add_to_transaction() marks the buf as
clean; if we intend to add a *particular* *snapshot* of the buffer's contents
to the transaction, it becomes important to also mark *that* particular snapshot
of buf as clean... otherwise, we will risk losing the dirty-bit information
corresponding to updates that occur after we took the snapshot. Hence this rules
out invoking add_to_transaction() with the write_lock_ dropped.

Put together, this means that we need to always call add_to_transaction() with
the buf's write_lock_ held, and mark the buf as clean before dropping the lock.