buf: Mark buf as clean before writing it back to disk or adding to transaction
Commit 73bb85ed3 "fs, buf: Fix the placement of mark_dirty() and mark_clean()"
changed the placement of mark_clean() in buf::writeback(). But it turns out
that the original placement was actually correct, and unfortunately that commit
failed to observe it, because it overlooked a scenario involving a write
operation running concurrently with the writeback.

Consider the following scenario:

        writeback                                        write
        ---------                                        -----

   acquire writeback_lock_

   read contents of the buf

   issue disk write

                                            acquire write_lock_ (via buf_writer)
                                            // automatically buf is marked dirty

                                            modify contents of the buf

   mark buf as clean


As seen from the above example, we lose the dirty bit from the fresh write,
because the mark_clean() inside writeback() overwrites it. Instead, if we had
marked the buf as clean in writeback() *before* reading the contents of buf,
then we wouldn't have this lost dirty-bit problem.

So revert back mark_clean() to its original location, and also do the same
update in add_to_transaction().