buf: Remove the 'buf' parameter from add_to_transaction()
Since add_to_transaction() is called with the buf's write_lock_ held,
we can just read the buf's contents and add it to the transaction. So
it is unnecessary to pass the buffer contents as an argument.

However, note that we can't issue a read() (like what ->writeback() does)
in order to read the block contents, because we are already holding the
seq-lock for write (and hence we will end up in a self-deadlock if we
invoke read() here).