buf, ide: Clarify the synchronous nature of the disk-write in writeback()
The call to idewrite() within writeback() is actually synchronous, because
idewrite() invokes idewait() at the end of the write operation. So the
comment about not having to wait for the write to finish isn't actually
true.

Furthermore, it is actually *necessary* (at least in the current design)
that the write be synchronous, lest we free up the buf (due to eviction)
mid-way through the copy operation, thus crashing the disk driver!

Clarify these aspects using appropriate comments inside writeback().