eviction: Introduce evict_pagecache() to evict pages from the page-cache
Add evict_pagecache() to go through each file in the filesystem that has been
referenced (by looking it up in the MFS layer) and drop all the cached pages
of those files. The helper named drop_pagecache() performs the actual task
of invoking mfile::put_page() on all the pages of the file.

We don't evict dirty pages, so as to be able to keep page-cache eviction
independent from sync/fsync.