eviction: Introduce mfile::put_page() to evict pages from the page-cache
We have mfile::get_page() to allocate page-cache pages and load file page
contents into them. In the future, we would like to have a mechanism to
evict selected pages from the page-cache. To achieve this, introduce the
counterpart of mfile::get_page() called mfile::put_page(), having a similar
interface.

Since the pages in the page-cache are managed using reference counts, we
just need to drop the reference count in mfile::put_page(), in order to evict
the page from the page-cache and free up its memory. However, there are a
few additional steps to be performed before dropping the reference count:

 - Once the page is evicted, a subsequent read() of the file must be able to
   detect the absence of the page and allocate a new one. This is achieved by
   clearing out the page_info pointer corresponding to this page.

 - This page might have been mapped by the page-tables of several processes.
   So we need to remove all those mappings before freeing the page. We use the
   rmap infrastructure to perform this efficiently.

Also, we take care to evict only clean pages from the page-cache (although
in principle, we could write-back dirty data pages of the file to the disk
and free up those pages as well, since these data writes don't have to be
wrapped inside a transaction).