page_info: Explicitly invoke page_info's destructor upon hitting onzero()
The page_info object is always allocated using placement new, so it is our
responsibility to invoke its destructor explicitly. So call the destructor
inside onzero(). Note that the destructor is responsible for freeing up the
memory allocated to rmap_pte; so this patch will fix that memory leak as well.