mnode: Add is_initialized() function to mnode
mnode::as_file() checks if the file has been initialized, and if not, it
automatically performs the initialization. But there are some cases where
we merely want to know if the file has been initialized or not (without
performing the actual initialization). So introduce is_initialized() for
use in those situations.