vm: Introduce vmap::clear_mapping() to clear out the page-mapping from vmdesc
Today vmap provides remove_mapping() to unmap a virtual page. But it also
unsets that page from the vpfs_ array, and hence it is suitable only for
scenarios involving destructive tear-down such as truncating a file.

For use-cases such as page-cache eviction, introduce a softer variant to
unmap a page, called clear_mapping(), which only clears the page-mapping
from the vmdesc, and doesn't unset it from vpfs_.

To further emphasize the distinction between these two cases, rename
the function remove_mapping() as delete_mapping().