oplog, page-eviction: Fix crash by clearing out dead objects from the cache
We were hitting a very weird bug that used to crash the kernel as shown below,
and was getting triggered only when running page-cache eviction along with
other tasks (which exercise the very common path: pagefault->ensure_page()->
add_pte()->add_mapping()->get_logger()):


evict_caches: dropping page-cache pages

kernel trap 14 err 0x0 cpu 0 cs 16 ds 24 ss 0
  rip ffffffffc015462a rsp ffffff001f9dfb60 rbp ffffff001f9dfbb0
  cr2 0000000000000050 cr3 000000001fa3a000 cr4 00000000000007b0
  rdi ffffff001fb9a640 rsi ffffffffc04a09d8 rdx 0000000000000000
  rcx 0000000000000002 r8  0000000000000019 r9  ffffffffc8a63900
  rax 0000000000000040 rbx ffffff001fb9a640 r10 0000000000000000
  r11 0000000000000040 r12 ffffff001fb9a650 r13 ffffffffc04a09c8
  r14 ffffff001fc8f680 r15 ffffffffc04a09d8 rflags 0000000000010002
  proc: name sh pid 63 kstack 0xffffff001f9d8000
  page fault: non-present page reading 0000000000000000 from kernel mode
  ffffffffc014d2ae
  ffffffffc01518d0
  ffffffffc0151d17
  ffffffffc01523f3
  ffffffffc0156bb4
  ffffffffc015762d
  ffffffffc0157fdd
  000000000040b8fa

The actual crash corresponded to this line in oplog's get_logger() code:

cur_obj->flush_logger(&my_way->logger_);

Apparently the virtual method flush_logger() couldn't get resolved and we
would end up dereferencing an invalid address.

However, this issue had seemingly no connection with page-cache eviction,
because it evicts only file-backed pages and doesn't touch any other pages
at all! Yet this bug was reproducible only with page-cache eviction in the
picture.

It turns out that the problem arises when an object that is logged using
oplog gets destroyed, but remains in the oplog cache. In such a scenario,
get_logger() finds a non-NULL pointer to that object in the cache, but since
it is already dead, any more operations on that object crash the kernel.

There are 2 users of oplog in the kernel: mfs-logical-log and rmap.
Objects created for maintaining mfs-logical-logs are never destroyed, and
hence this code path is never executed. As for rmap (which is a part of
page_info), they are dynamically allocated in the page_info constructor,
but since page_info itself is allocated using placement new, its destructor
was never getting invoked.

While implementing page-cache eviction, this bug was noticed and fixed by
commit 5a7667f6621 (page_info: Explicitly invoke page_info's destructor upon
hitting onzero()). From then on, whenever a page is evicted, its page_info
destructor would get invoked, which deleted the rmap object, leading to the
invocation of clear_loggers() (through tsc_logged_object's destructor).

While clear_loggers() does reset the object's loggers, it fails to delete
the object from the oplog cache! Hence, get_logger() would end up finding
this stale (dead) object in the cache and hence crash the kernel, trying to
invoke flush_logger() on that object.

So fix this by deleting the pointer to the object from the cache whenever
we are done processing its logged operations, and especially when the object
is about to get destroyed. This ensures that we don't leave any stale/invalid
entries behind.