Implement AHCI disk whitelisting
The sv6 AHCI driver will only attach to disks whose model and serial
numbers are white-listed in kernel/ahci.cc.  This is intended to prevent
corruption to existing disks on josmp, tom, and ben.  Currently the only
white-listed disk is QEMU's.  Once we figure out what disk we want to
run experiments on, we should add it to kernel/ahci.cc.