benchmarks, bin: Import the Sprite LFS Largefile benchmark
This benchmark was described in the SOSP (1991) paper titled "The Design and
Implementation of a Log-Structured File System" by Mendel Rosenblum and
John K. Ousterhout.

Source code for the benchmark was imported from
http://fsbench.filesystems.org/bench/sprite-lfs.tar.gz

Haven't added it to the Makefile yet, because it needs some fixes to build on
sv6. Also, I have renamed the file as lfs-largefile.c