benchmarks, bin: Import the Sprite LFS Smallfile benchmark
This benchmark was described in the SOSP (1991) paper titled "The Design and
Implementation of a Log-Structured File System" by Mendel Rosenblum and
John K. Ousterhout.

Source code for the benchmark was imported from
http://www.fsl.cs.sunysb.edu/docs/fsbench/lfs.tar.gz

(This code was described/implemented in conjunction with the paper titled
"A Nine Year Study of File System and Storage Benchmarking", by Avishay
Traeger et al. A copy of the paper can be found at:
http://www.fsl.cs.sunysb.edu/docs/fsbench/fsbench.pdf)

Haven't added it to the Makefile yet, because it needs some fixes to build on
sv6. Also, I have renamed the file as lfs-smallfile.c