scalefs: Separate unlink and delete by adding a new transaction to delete a file
According to POSIX, an inode (and its file-contents) can be deleted only
after the last link to that inode has been removed *and* the last open file
descriptor referring to that file has been closed. At the moment, sv6/scalefs
doesn't handle the latter case correctly: it deletes the file as soon as the
file's last link has been dropped. A simple example where sv6 fails is:

  open();
  unlink(); // sv6/scalefs deletes the file here.

  // write(); or read();

  close(); // The file should not be deleted until this step.

To fix this, introduce a new transaction to delete the file. This will help us
in deferring the deletion of the file after unlink, until its last open file
descriptor has been closed.