scalefs: Defer file delete until its mnode's reference count drops to zero
POSIX specifies that an inode and its file contents should be kept around
until its last link has been removed *and* its last open file descriptor has
been closed. At the moment, sv6/scalefs deletes the file as soon as its link
count drops to zero, which is incorrect.

A file's mnode reference count drops to zero only when it satisfies both the
above requirements : link-count == 0 and no open file descriptors. So arrange
to delete the file when the mnode's onzero() method is invoked. Also, use the
newly introduced "delete" transaction to delete the file, because file deletion
has to be done within a transaction, to avoid its blocks from being reused
prematurely.

The underlying inode tracks only the number of links it has, and hence has
no notion of whether there are any open file descriptors referring to it (the
mnode tracks this latter aspect as well). So, perform an extra increment of
the reference count of the inode when initializing it, so that it can be
dropped when the mnode's reference count hits zero. Thus we make use of the
mnode's tracking facilities to manage the life-time of the underlying inode.

This design doesn't yet handle the scenario where the file has an open file
descriptor and its last link drops to zero, but the kernel crashes before
the file could be closed. In that case, the unlink operation would have been
logged/applied, but the corresponding delete operation would be lost. Hence,
the file's blocks will remain unreclaimable upon reboot, even though there
are no links to its inode.