benchmarks, LFS Smallfile: Spread files among multiple directories
This particular implementation of the LFS Smallfile benchmark creates all the
files (10000 of them!) in the same directory. So, the performance of the
benchmark can be easily dominated by the time spent in searching for the
directory entries. It would be better to spread the files among several
directories to reduce this undesirable effect. So add a "-s" parameter to
specify the spread, indicating how many directories we want the files to be
spread amongst.

Also, the benchmark is quite inflexible as it is, because it hard-codes the
number of files. So add a switch to customize this as well.