fs, scalefs: Don't add an extra link count to the inode when creating directories
create_dir_if_new() creates a new directory and adds its parent directory
("..") as its first directory entry. But while doing that, it also
inadvertently adds an additional link count to the newly created directory,
by passing 'true' as the last argument to dirlink(). Thus the directory will
be created with a link count of 2, instead of 1. As a result, when the
directory is unlinked, it won't actually be freed from the disk, because its
inode link count would not have dropped to zero.

Fix this by passing 'false' as the last argument to dirlink() in this case.